style : 채팅목록, 채팅방 repository 분리
